blog app3/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-filter
